test: DB接続とマイグレーションをパッケージ全体で1回だけ実行してテストを高速化
TestMain で sharedTestDB を初期化し、各テスト関数での rdb.Migrate 呼び出しを廃止。
これによりスキーマ照会のオーバーヘッドがパッケージあたり1回に削減される。
